+++ src/main/kotlin/com/github/jensim/megamanipulator/onboarding/OnboardingOperator.kt
@@ -1,11 +1,17 @@
 package com.github.jensim.megamanipulator.onboarding
 
+import com.github.jensim.megamanipulator.project.MegaManipulatorSettingsState.Companion.seenOnBoarding
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.ui.popup.Balloon
 import com.intellij.openapi.ui.popup.JBPopupFactory
 import com.intellij.openapi.wm.WindowManager
 import com.intellij.ui.awt.RelativePoint
+import com.intellij.ui.layout.panel
 import java.util.EnumMap
+import javax.swing.JButton
 import javax.swing.JComponent
+import javax.swing.JLayeredPane
+import javax.swing.JTextArea
 
 object OnboardingOperator {
 
@@ -21,24 +27,42 @@ object OnboardingOperator {
     }
 
     fun display(id: OnboardingId) {
-        val targetComponent = reg[id]
-        val popupFactory: JBPopupFactory = JBPopupFactory.getInstance()
-        val popupLocation: RelativePoint = targetComponent
-            ?.let { popupFactory.guessBestPopupLocation(targetComponent) }
-            ?: try {
-                getWindowComponent()?.let { popupFactory.guessBestPopupLocation(it) }
-            } catch (e: Exception) {
-                null
-            }
-            ?: return
+        if (seenOnBoarding(id)) {
+            return
+        }
 
-        val popup = popupFactory.createMessage(id.text)
-        popup.setCaption(id.title)
-        popup.showInFocusCenter()
-        popup.show(popupLocation)
+        val popupFactory: JBPopupFactory = JBPopupFactory.getInstance()
+        val popupLocation: RelativePoint? = reg[id]?.let { popupFactory.guessBestPopupLocation(it) }
+
+        val closeButton = JButton("Ok")
+        val panel = panel {
+            row {
+                component(JTextArea(id.text).apply { })
+            }
+            row {
+                component(closeButton)
+            }
+        }
+        val balloon = popupFactory.createDialogBalloonBuilder(panel, id.title)
+            .createBalloon()
+
+        closeButton.addActionListener {
+            balloon.hide(true)
+            id.next?.let {
+                display(it)
+            }
+        }
+        balloon.setAnimationEnabled(true)
+        if (popupLocation != null) {
+            balloon.show(popupLocation, Balloon.Position.above)
+        } else {
+            getWindowComponent2()?.let {
+                balloon.showInCenterOf(it)
+            }
+        }
     }
 
-    private fun getWindowComponent(): JComponent? {
-        return WindowManager.getInstance().getIdeFrame(project)?.component
+    private fun getWindowComponent2(): JLayeredPane? {
+        return WindowManager.getInstance().getFrame(project)?.layeredPane
     }
 }